6420 - Move wagtail hooks into wagtail_hooks.py file
