+++ network-api/networkapi/wagtailpages/wagtail_hooks.py
@@ -4,7 +4,11 @@
 
 from django.templatetags.static import static
 from django.core.cache import cache
+from django.urls import reverse
+from django.utils.html import format_html
 
+from wagtail.admin.menu import MenuItem
+from django.utils.translation import ugettext_lazy as _
 import wagtail.admin.rich_text.editors.draftail.features as draftail_features
 from wagtail.admin.rich_text.converters.html_to_contentstate import InlineStyleElementHandler
 from wagtail.core import hooks
@@ -127,3 +131,21 @@ def global_admin_js():
 def global_admin_css():
     max_length_css = static('wagtailadmin/css/max-length-field.css')
     return f'<link rel="stylesheet" href="{max_length_css}">'
+
+
+class HowToWagtailMenuItem(MenuItem):
+    def is_shown(self, request):
+        return True
+
+
+@hooks.register('register_admin_menu_item')
+def register_howto_menu_item():
+    return HowToWagtailMenuItem(
+        _('How Do I Wagtail'), reverse('how-do-i-wagtail'),
+        name='howdoIwagtail', classnames='icon icon-help', order=900
+    )
+
+@hooks.register("insert_global_admin_css", order=100)
+def global_admin_css():
+    """Add /static/css/admin.css to the admin."""
+    return format_html('<link rel="stylesheet" href="{}">', static("css/admin.css"))
